requirements.txt
pyyaml>=6.0
matplotlib>=3.5.0